根据功能重新组织 sniper rpc 源码
