+++ cmd/sniper/rpc/server.go
@@ -17,7 +17,11 @@ import (
 	"golang.org/x/tools/go/ast/astutil"
 )
 
+var serverFile string
+
 func genOrUpdateServer() {
+	serverFile = fmt.Sprintf("rpc/%s/v%s/%s.go", server, version, service)
+
 	if !fileExists(serverFile) {
 		tpl := &srvTpl{
 			Server:  server,
@@ -275,15 +279,4 @@ func scanDefinedFuncs(file string) map[string]bool {
 	}
 
 	return fs
-}
-
-// 判断文件是否存在
-func fileExists(file string) bool {
-	fd, err := os.Open(file)
-	defer fd.Close()
-
-	if err != nil && os.IsNotExist(err) {
-		return false
-	}
-	return true
 }
